Redis on VM statements in benchmark pages updated.
While it is true you lose a percentage of performance when running using
virtualization, the reality is that:

1) To point at exact figures may be misleading because it changes from
environment to environment.
2) Usually people want to run Redis virutalized using "cloud providers"
so the point is if it runs well or not in VMs, and it works well. A
comparison between a server and a Linode or EC2 instance is basically
pointless.
3) The real problems people have, for example using EC2, are mostly
related to things like horrible non-local disk latency, not
virtualization per se.

However the general idea that bare metal is better if you have a choice
remains in the documentation, because is a good one, even if it is
pretty obvious with any database system.