Remove Commons Clause modules
Hi, all. Vice President of the Open Source Initiative (OSI) here.

The Commons Clause, when applied to an OSI-approved license, creates a new license that violates the Open Source Definition (OSD). The OSI does not approve licenses that violate the OSD and does not approve of the Commons Clause in general. Therefore modules with licenses that include the Commons Clause are not OSI-approved.

As this page says it lists only OSI-approved licenses, I've removed the Commons Clause-using modules from the list so that the page is now accurate.