Update project url of the radix go redis client (#1037)
After a few years of design and testing, with lots of feedback from
people testing it, I'm ready to say that the newest version of radix is
ready to be used. This version is faster and more flexible, and is
better able to handle new features like streams and RESP3.

Due to go's new module dependency system, this will be the last time the
url of the project needs to change due to a major version change.