fix minor error in eval doc
The section 'Using Lua scripting in RESP3 mode' includes the following sentence:

The RESP3 protocol is semantically more powerful, however most scripts are ok with using just RESP3.

However, the sentence makes no sense. I suspect that it was a typo, which changes the meaning though.
The sentence is probably supposed to go like this:

The RESP3 protocol is semantically more powerful, however most scripts are ok with using just RESP2.