Remove <CR> from protocol as it doesn't work
I have tested this in redis in a linux server, and adding <CR> in the
protocol file to pipe to redis-cli --pipe just doesn't work. It only
works with "normal" line endings of <LF> only.

However the examples with printf require \r to work. Why the behavior
of redis-cli --pipe is different?