Remove unnecessary code in streams consumer implementation
When the loop is still going through backlog, whether id is '0-0' or the id of the last processed message, both have same outcome in this context. 
Also, once the consumer is caught up with backlog, the id used becomes ``>``. So it looks like saving the id of last processed message is unnecessary.